requirements.txt
python-dotenv
openai
pinecone
fastapi
uvicorn
